update graphql-kotlin implementation with latest schema changes
Update to expected schema that verifies `@requires` directive functionality.

Changes:

* split code into separate class for easier maintainability
* add missing `ProductDimension.unit` value
* add `averageProductsCreatedPerYear: Int @requires(fields: "yearsOfEmployment")` on `User` type
* add new `yearsOfEmployment: Int! @external` field on `User` type`

Related Issues:

* https://github.com/apollographql/apollo-federation-subgraph-compatibility/issues/128
