+++ implementations/graphql-kotlin/src/main/kotlin/com/example/demo/ProductQuery.kt
@@ -1,0 +1,11 @@
+package com.example.demo
+
+import com.example.demo.model.Product
+import com.expediagroup.graphql.generator.scalars.ID
+import com.expediagroup.graphql.server.operations.Query
+import org.springframework.stereotype.Component
+
+@Component
+class ProductQuery : Query {
+    fun product(id: ID) = Product.byID(id)
+}
