+++ implementations/graphql-kotlin/src/main/kotlin/com/example/demo/model/ProductDimension.kt
@@ -1,0 +1,19 @@
+package com.example.demo.model
+
+import com.expediagroup.graphql.generator.federation.directives.InaccessibleDirective
+import com.expediagroup.graphql.generator.federation.directives.ShareableDirective
+
+/*
+type ProductDimension @shareable {
+  size: String
+  weight: Float
+  unit: String @inaccessible
+}
+ */
+@ShareableDirective
+data class ProductDimension(
+    val size: String? = null,
+    val weight: Float? = null,
+    @InaccessibleDirective
+    val unit: String? = null
+)
