+++ implementations/graphql-kotlin/src/main/kotlin/com/example/demo/model/ProductVariation.kt
@@ -1,0 +1,12 @@
+package com.example.demo.model
+
+import com.expediagroup.graphql.generator.scalars.ID
+
+/*
+type ProductVariation {
+  id: ID!
+}
+ */
+data class ProductVariation(
+    val id: ID
+)
